requirements.txt
certifi==2020.12.5
chardet==4.0.0
Faker==5.7.0
idna==2.10
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
text-unidecode==1.3
urllib3==1.26.2
